More fixes to migrate to Todoist v9 Sync API.
This prorgam was failing again, and it appears that this time it is because
there are more changes in the Todoist v9 Sync API that it initially appeared:
* GET requests are now POST
* data must be passed as FormData instead of QueryString
* The token must be passed in 'Authorization: Bearer ...' instead of data.

For some reason this wasn't necessary 1 week ago, I guess they still kept some
v8 API code which is now removed or something. Anyway, use the API properly.